requirements.txt
pandas
pytest
streamlit
requests
numpy
google-genai
